refactor: streamline core
