+++ packages/fiber/src/core/loop.ts
@@ -1,9 +1,10 @@
-import * as THREE from 'three'
-import { Root } from './renderer'
-import { RootState, Subscription } from './store'
+import { _roots } from './renderer'
+import type { RootState } from './store'
 
 export type GlobalRenderCallback = (timestamp: number) => void
-type SubItem = { callback: GlobalRenderCallback }
+interface SubItem {
+  callback: GlobalRenderCallback
+}
 
 function createSubs(callback: GlobalRenderCallback, subs: Set<SubItem>): () => void {
   const sub = { callback }
@@ -11,10 +12,9 @@ function createSubs(callback: GlobalRenderCallback, subs: Set<SubItem>): () => v
   return () => void subs.delete(sub)
 }
 
-let i
-let globalEffects: Set<SubItem> = new Set()
-let globalAfterEffects: Set<SubItem> = new Set()
-let globalTailEffects: Set<SubItem> = new Set()
+const globalEffects = new Set<SubItem>()
+const globalAfterEffects = new Set<SubItem>()
+const globalTailEffects = new Set<SubItem>()
 
 /**
  * Adds a global render callback which is called each frame.
@@ -54,8 +54,6 @@ export function flushGlobalEffects(type: GlobalEffectType, timestamp: number): v
   }
 }
 
-let subscribers: Subscription[]
-let subscription: Subscription
 function update(timestamp: number, state: RootState, frame?: XRFrame) {
   // Run local effects
   let delta = state.clock.getDelta()
@@ -76,78 +74,70 @@ function update(timestamp: number, state: RootState, frame?: XRFrame) {
   return state.frameloop === 'always' ? 1 : state.internal.frames
 }
 
-export function createLoop<TCanvas>(roots: Map<TCanvas, Root>) {
-  let running = false
-  let repeat: number
-  let frame: number
-  let state: RootState
+let running = false
+let repeat: number
+let frame: number
+let state: RootState
 
-  function loop(timestamp: number): void {
-    frame = requestAnimationFrame(loop)
+export function loop(timestamp: number): void {
+  frame = requestAnimationFrame(loop)
+  running = true
+  repeat = 0
+
+  // Run effects
+  flushGlobalEffects('before', timestamp)
+
+  // Render all roots
+  for (const root of _roots.values()) {
+    state = root.store.getState()
+
+    // If the frameloop is invalidated, do not run another frame
+    if (
+      state.internal.active &&
+      (state.frameloop === 'always' || state.internal.frames > 0) &&
+      !state.gl.xr?.isPresenting
+    ) {
+      repeat += update(timestamp, state)
+    }
+  }
+
+  // Run after-effects
+  flushGlobalEffects('after', timestamp)
+
+  // Stop the loop if nothing invalidates it
+  if (repeat === 0) {
+    // Tail call effects, they are called when rendering stops
+    flushGlobalEffects('tail', timestamp)
+
+    // Flag end of operation
+    running = false
+    return cancelAnimationFrame(frame)
+  }
+}
+
+/**
+ * Invalidates the view, requesting a frame to be rendered. Will globally invalidate unless passed a root's state.
+ * @see https://docs.pmnd.rs/react-three-fiber/api/additional-exports#invalidate
+ */
+export function invalidate(state?: RootState, frames = 1): void {
+  if (!state) return _roots.forEach((root) => invalidate(root.store.getState()), frames)
+  if (state.gl.xr?.isPresenting || !state.internal.active || state.frameloop === 'never') return
+  // Increase frames, do not go higher than 60
+  state.internal.frames = Math.min(60, state.internal.frames + frames)
+  // If the render-loop isn't active, start it
+  if (!running) {
     running = true
-    repeat = 0
-
-    // Run effects
-    flushGlobalEffects('before', timestamp)
-
-    // Render all roots
-    for (const root of roots.values()) {
-      state = root.store.getState()
-
-      // If the frameloop is invalidated, do not run another frame
-      if (
-        state.internal.active &&
-        (state.frameloop === 'always' || state.internal.frames > 0) &&
-        !state.gl.xr?.isPresenting
-      ) {
-        repeat += update(timestamp, state)
-      }
-    }
-
-    // Run after-effects
-    flushGlobalEffects('after', timestamp)
-
-    // Stop the loop if nothing invalidates it
-    if (repeat === 0) {
-      // Tail call effects, they are called when rendering stops
-      flushGlobalEffects('tail', timestamp)
-
-      // Flag end of operation
-      running = false
-      return cancelAnimationFrame(frame)
-    }
+    requestAnimationFrame(loop)
   }
+}
 
-  function invalidate(state?: RootState, frames = 1): void {
-    if (!state) return roots.forEach((root) => invalidate(root.store.getState()), frames)
-    if (state.gl.xr?.isPresenting || !state.internal.active || state.frameloop === 'never') return
-    // Increase frames, do not go higher than 60
-    state.internal.frames = Math.min(60, state.internal.frames + frames)
-    // If the render-loop isn't active, start it
-    if (!running) {
-      running = true
-      requestAnimationFrame(loop)
-    }
-  }
-
-  function advance(timestamp: number, runGlobalEffects: boolean = true, state?: RootState, frame?: XRFrame): void {
-    if (runGlobalEffects) flushGlobalEffects('before', timestamp)
-    if (!state) for (const root of roots.values()) update(timestamp, root.store.getState())
-    else update(timestamp, state, frame)
-    if (runGlobalEffects) flushGlobalEffects('after', timestamp)
-  }
-
-  return {
-    loop,
-    /**
-     * Invalidates the view, requesting a frame to be rendered. Will globally invalidate unless passed a root's state.
-     * @see https://docs.pmnd.rs/react-three-fiber/api/additional-exports#invalidate
-     */
-    invalidate,
-    /**
-     * Advances the frameloop and runs render effects, useful for when manually rendering via `frameloop="never"`.
-     * @see https://docs.pmnd.rs/react-three-fiber/api/additional-exports#advance
-     */
-    advance,
-  }
+/**
+ * Advances the frameloop and runs render effects, useful for when manually rendering via `frameloop="never"`.
+ * @see https://docs.pmnd.rs/react-three-fiber/api/additional-exports#advance
+ */
+export function advance(timestamp: number, runGlobalEffects: boolean = true, state?: RootState, frame?: XRFrame): void {
+  if (runGlobalEffects) flushGlobalEffects('before', timestamp)
+  if (!state) for (const root of _roots.values()) update(timestamp, root.store.getState())
+  else update(timestamp, state, frame)
+  if (runGlobalEffects) flushGlobalEffects('after', timestamp)
 }
